refactor: 라벨링 파트 파일 이동
