+++ src/app/label/[imageKey]/_components/label-step-layout.tsx
