+++ src/app/label/[imageKey]/_components/label-step/index.ts
