+++ src/app/label/[imageKey]/_components/label-step/label-step3.tsx
@@ -6,8 +6,8 @@ import useUpdateLabel from '@/hooks/use-update-label';
 import { TOTAL_LABELING_STEPS } from '@/lib/constants';
 import { WidthClass } from '@/types/api/label';
 
-import { labelOption } from '../../types/label-option';
-import { LabelStepProps } from '../../types/label-step';
+import { labelOption } from '../../_types/label-option';
+import { LabelStepProps } from '../../_types/label-step';
 import LabelStepLayout from '../label-step-layout';
 
 const LABEL_STEP_3_OPTIONS: labelOption<WidthClass>[] = [
