+++ src/app/label/[imageKey]/_types/label-option.ts
