+++ src/app/label/[imageKey]/_types/label-step.ts
